Slicer Custom Application Template 2.0
This project is the version 2.0 of a template we have been using
at Kitware to create both commercial and open-source Slicer based
applications since August 2013.

It exists only because of the skilled, collaborative and awesome communities
maintaining the projects it depends on.

The original template was primarily developed by the following
contributors:

Co-authored-by: Julien Finet <julien.finet@kitware.com>
Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>
Co-authored-by: Dzenan Zukic <dzenan.zukic@kitware.com>
Co-authored-by: Johan Andruejol <johan.andruejol@kitware.com>
Co-authored-by: Alexis Girault <alexis.girault@kitware.com>
Co-authored-by: Cory Quammen <cory.quammen@kitware.com>
Co-authored-by: Jamie Snape <jamie.snape@kitware.com>